chore: shift from settingsv2 to agency
